% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_asserts.R
\name{assert_anc_client_numbers}
\alias{assert_anc_client_numbers}
\title{Provides some checks on the validity of the ANC data}
\usage{
assert_anc_client_numbers(data)
}
\arguments{
\item{data}{ANC data to validate}
}
\value{
TRUE if data passes validity checks, else throws error
}
\description{
Provides some checks on the validity of the ANC data
}
\keyword{internal}
