% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_asserts.R
\name{assert_column_positive_numeric}
\alias{assert_column_positive_numeric}
\title{Check that the values of a column are positive numeric only}
\usage{
assert_column_positive_numeric(data, column_names)
}
\arguments{
\item{data}{data to be checked}

\item{column_names}{column name to be checked}
}
\value{
TRUE if numeric positive else throws error
}
\description{
Check that the values of a column are positive numeric only
}
\keyword{internal}
