% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_asserts.R
\name{assert_expected_values}
\alias{assert_expected_values}
\title{Check that values for specified column belong in a set of expected values}
\usage{
assert_expected_values(data, column_name, expected_values, all_values = FALSE)
}
\arguments{
\item{data}{The data to check}

\item{column_name}{The column name to check}

\item{expected_values}{The set of expected values}

\item{all_values}{If TRUE check that the column contains all expected values}
}
\value{
TRUE is valid otherwise throws an error
}
\description{
Optionally check that the column contains all expected values.
}
\keyword{internal}
