% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_asserts.R
\name{assert_no_na}
\alias{assert_no_na}
\title{Check that the values of a column are not NA}
\usage{
assert_no_na(data, column_name)
}
\arguments{
\item{data}{data to be checked}

\item{column_names}{column name to be checked}
}
\value{
TRUE if column contains no NA values else throws error
}
\description{
Check that the values of a column are not NA
}
\keyword{internal}
