% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_asserts.R
\name{assert_single_source}
\alias{assert_single_source}
\title{Checks that the data source column contains a single value}
\usage{
assert_single_source(data)
}
\arguments{
\item{data}{Data to check source column for single value}
}
\value{
TRUE is valid else throws error
}
\description{
Checks that the data source column contains a single value
}
\keyword{internal}
