% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_asserts.R
\name{assert_unique_combinations}
\alias{assert_unique_combinations}
\title{Check for unique combinations of values in each row of selected columns}
\usage{
assert_unique_combinations(data, columns_for_unique)
}
\arguments{
\item{columns_for_unique}{}
}
\value{
TRUE if the required combinations are unique, else throws error
}
\description{
Check for unique combinations of values in each row of selected columns
}
\keyword{internal}
