% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_helpers.R
\name{build_json}
\alias{build_json}
\title{Build JSON from template and a set of params}
\usage{
build_json(template, params)
}
\arguments{
\item{template}{Template JSON of payload}

\item{params}{List of named key value pairs for substituting from template.}
}
\value{
JSON built from template and params.
}
\description{
Build JSON from template and a set of params
}
\keyword{internal}
