% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse}
\alias{collapse}
\title{Collapse a vector into a human readable format for returning in messages}
\usage{
collapse(vector, collapse = ", ", limit = 10, end = ", ...")
}
\arguments{
\item{vector}{The vector to collapse}

\item{collapse}{Separator between each element after collapse}

\item{limit}{Max number of items from vector to return in string}

\item{end}{If length of vector over the max a string to append to the end}
}
\value{
The collapsed vector as a string.
}
\description{
Collapse a vector into a human readable format for returning in messages
}
\keyword{internal}
