% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{construct_tree}
\alias{construct_tree}
\title{Create an ordered tree from a data frame.}
\usage{
construct_tree(
  data,
  id_column = 1,
  parent_id_column = 2,
  sort_order_column = 3
)
}
\arguments{
\item{data}{Data frame with at minimum id, parent id and sort order columns
to construct tree from.}

\item{id_column}{Index of column containing IDs.}

\item{parent_id_column}{Index of column containing parent IDs.}

\item{sort_order_column}{Index of column containing sort orders.}
}
\value{
The tree represented as a list.
}
\description{
Will create a tree as a nested list from a data frame with specified order.
}
\keyword{internal}
