% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{do_endpoint_model_options}
\alias{do_endpoint_model_options}
\title{Get model options declarative UI from naomi and input data}
\usage{
do_endpoint_model_options(shape, survey, programme, anc)
}
\arguments{
\item{shape}{A file object (path, hash, filename) corresponding to
the input shape file}

\item{survey}{Path to survey file}

\item{programme}{Path to optional programme file}

\item{anc}{Path to optional ANC file}
}
\value{
The model options declarative JSON UI.
}
\description{
Get's UI template from Naomi and substitutes any params within the
template with real values.
}
\keyword{internal}
