% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{do_validate_anc}
\alias{do_validate_anc}
\title{Validate ANC data file.}
\usage{
do_validate_anc(anc)
}
\arguments{
\item{anc}{Path to input population file.}
}
\value{
An error if invalid.
}
\description{
Check that ANC data file can be read and return serialised data.
}
\keyword{internal}
