% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{do_validate_model_options}
\alias{do_validate_model_options}
\title{Validate model options}
\usage{
do_validate_model_options(data, options)
}
\arguments{
\item{data}{The set of input data for the model run}

\item{options}{Key value list of model options}
}
\value{
TRUE if valid else throws an error
}
\description{
Validate model options
}
\keyword{internal}
