% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{do_validate_population}
\alias{do_validate_population}
\title{Validate population file.}
\usage{
do_validate_population(population)
}
\arguments{
\item{population}{Path to input population file.}
}
\value{
An error if invalid.
}
\description{
Check that population file can be read.
}
\keyword{internal}
