% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{do_validate_programme}
\alias{do_validate_programme}
\title{Validate programme ART data file.}
\usage{
do_validate_programme(programme, shape)
}
\arguments{
\item{programme}{A file object (path, hash, filename) corresponding to
the input population file.}

\item{shape}{A file object (path, hash, filename) corresponding to
the input shape file.}
}
\value{
An error if invalid.
}
\description{
Check that programme ART data file can be read and return serialised data.
}
\keyword{internal}
