% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{do_validate_shape}
\alias{do_validate_shape}
\title{Validate shape file and return geojson for plotting.}
\usage{
do_validate_shape(shape)
}
\arguments{
\item{shape}{A file object (path, hash, filename) corresponding to
the input shape file.}
}
\value{
An error if invalid or the geojson if valid.
}
\description{
This checks that the geojson is for a single country and that each
feature has an area id associated with it.
}
\keyword{internal}
