% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{do_validate_vmmc}
\alias{do_validate_vmmc}
\title{Validate VMMC data file.}
\usage{
do_validate_vmmc(vmmc, shape, strict = TRUE)
}
\arguments{
\item{vmmc}{A file object (path, hash, filename) corresponding to
the input VMMC file.}

\item{shape}{A file object (path, hash, filename) corresponding to
the input shape file.}

\item{strict}{If FALSE then run less stringent validation rules, used
for data exploration mode.}
}
\value{
An error if invalid.
}
\description{
Check that VMMC data file can be read.
}
\keyword{internal}
