% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{endpoint_download_summary}
\alias{endpoint_download_summary}
\title{Download summary zip file.}
\usage{
endpoint_download_summary(queue)
}
\arguments{
\item{queue}{The queue used to retrieve download from task id.}
}
\value{
Bytes of zip file.
}
\description{
Returns a function which returns bytes of zip.
}
\keyword{internal}
