% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{endpoint_model_options}
\alias{endpoint_model_options}
\title{Get function to generate model options from Naomi template and input files}
\usage{
endpoint_model_options(options_template)
}
\arguments{
\item{options_template}{The naomi options template.}
}
\value{
Function to generate model options from input data.
}
\description{
This is used to build the model run options UI in the front end.
}
\keyword{internal}
