% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{endpoint_model_options_validate}
\alias{endpoint_model_options_validate}
\title{Validate that set of model run options is okay to submit for a model run}
\usage{
endpoint_model_options_validate(req, res, data, options)
}
\arguments{
\item{req}{The request object}

\item{res}{The response object}

\item{data}{The set of input data for the model run}

\item{options}{key-value list of selected options}
}
\value{
Response with data valid if true and errors if not valid
}
\description{
Validate that set of model run options is okay to submit for a model run
}
\keyword{internal}
