% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{endpoint_validate_baseline_combined}
\alias{endpoint_validate_baseline_combined}
\title{Validate the baseline data as a collection.}
\usage{
endpoint_validate_baseline_combined(req, res, pjnz, shape, population)
}
\arguments{
\item{req}{The request as PlumberRequest object.}

\item{res}{The response as a PlumberResponse object.}

\item{pjnz}{Path to the pjnz file.}

\item{shape}{Path to the shape file.}

\item{population}{Path to the population file.}
}
\value{
Validated JSON response with data and incidcation of success.
}
\description{
Checks whether the collection of baseline data (pjnz, shape and population)
are consistent.
}
\keyword{internal}
