% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{endpoint_validate_input}
\alias{endpoint_validate_input}
\title{Validate an input file and return an indication of success and
if successful return the data required by UI.}
\usage{
endpoint_validate_input(req, res, type, path)
}
\arguments{
\item{req}{The request as PlumberRequest object.}

\item{res}{The response as a PlumberResponse object.}

\item{type}{The type of file to validate: pjnz, shape, population, ANC,
survey or programme.}

\item{path}{Path to the file to validate.}
}
\value{
Validates JSON response with data and incidcation of success.
}
\description{
Validate an input file and return an indication of success and
if successful return the data required by UI.
}
\keyword{internal}
