% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{endpoint_validate_survey_programme}
\alias{endpoint_validate_survey_programme}
\title{Validate survey and programme data including consistency with shape file
returning an indication of success and if successful return the data
required by UI.}
\usage{
endpoint_validate_survey_programme(req, res, type, path, shape)
}
\arguments{
\item{req}{The request as PlumberRequest object.}

\item{res}{The response as a PlumberResponse object.}

\item{type}{The type of file to validate: ANC, survey or programme.}

\item{path}{Path to the file to validate.}

\item{shape}{Path to shape file for comparison.}
}
\value{
Validated JSON response with data and incidcation of success.
}
\description{
Validate survey and programme data including consistency with shape file
returning an indication of success and if successful return the data
required by UI.
}
\keyword{internal}
