% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_notes}
\alias{format_notes}
\title{Take "notes" from DownloadSubmitRequest schema and transform
into single string for passing to naomi}
\usage{
format_notes(notes)
}
\arguments{
\item{notes}{The notes to parse}
}
\value{
The notes as a single string.
}
\description{
Take "notes" from DownloadSubmitRequest schema and transform
into single string for passing to naomi
}
\keyword{internal}
