% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_hint_id}
\alias{get_hint_id}
\title{Get hint ID from naomi ID}
\usage{
get_hint_id(naomi_id)
}
\arguments{
\item{naomi_id}{The ID used in Naomi to identify an indicator.}
}
\value{
The ID as used in hint.
}
\description{
Get hint ID from naomi ID
}
\keyword{internal}
