% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{get_indicator_data}
\alias{get_indicator_data}
\title{Filter data to only return rows containing data for the specified indicator}
\usage{
get_indicator_data(file, type, indicator)
}
\arguments{
\item{type}{The type of input data.}

\item{indicator}{The indicator to filter on.}

\item{data}{The input data type to filter.}
}
\value{
The read data filtered for the indicator
}
\description{
This filters data based on metadata. If wide format data it will
return the input with no filtering. If long format then it will only
return the rows containing the specified indicator.
}
\keyword{internal}
