% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{get_json_regions}
\alias{get_json_regions}
\title{Get regions from geoJSON}
\usage{
get_json_regions(json)
}
\arguments{
\item{json}{The Geojson.}
}
\value{
The list of regions from the geojson.
}
\description{
Get regions from geoJSON
}
\keyword{internal}
