% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_selected_filter_options}
\alias{get_selected_filter_options}
\title{Get selected filter options from full list of output filters.}
\usage{
get_selected_filter_options(output_filters, filter_type, ids = NULL)
}
\arguments{
\item{output_filters}{The full set of output filters}

\item{filter_type}{The type of filter to get options for, area, age, sex, quarter}

\item{ids}{Set of option IDs to reutrn, or NULL for all options}
}
\value{
The selected filter options
}
\description{
Gets the filter options of a particular type matching a set of IDs.
If ids are NULL then this returns all options.
}
\keyword{internal}
