% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_selected_mappings}
\alias{get_selected_mappings}
\title{Get selected id-label mapping from list of filter options or column mappings}
\usage{
get_selected_mappings(mappings, type, ids = NULL, key = "options")
}
\arguments{
\item{mappings}{The full set of output filters or column mappings}

\item{type}{The type of column to get mappings for, area, age, sex, quarter}

\item{ids}{Set of IDs to return, or NULL for all options}

\item{key}{The key to use for individual options. For filter mappings this
is "options" for column mapping this is "value"}
}
\value{
The selected mappings
}
\description{
Gets the id to label mapping of a particular type matching a set of IDs.
If ids are NULL then this returns all options.
}
\keyword{internal}
