% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerun.R
\name{prerun_push}
\alias{prerun_push}
\title{Push prerun model results to naomi.dide.ic.ac.uk}
\usage{
prerun_push(path, model_output = "model-output.qs")
}
\arguments{
\item{path}{Path to your results}

\item{model_output}{Path, within \code{path} to your model output qs}
}
\description{
Push prerun model results to naomi.dide.ic.ac.uk
}
