% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{read_long_indicator_filters}
\alias{read_long_indicator_filters}
\title{Read filters from long format data}
\usage{
read_long_indicator_filters(data, type)
}
\arguments{
\item{data}{Long format data}

\item{type}{The type of data set}
}
\value{
Indicator filters
}
\description{
Expect input data with headers like
}
\details{
x, y, z, indicator, value

Where indicator columns describes the type of indicator the value is for.
For the data type we have look for the list of possible indicators from the
metadata. If that indicator exists in the metadata return it as a possible
filter with ID and name.
}
\keyword{internal}
