% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{read_wide_indicator_filters}
\alias{read_wide_indicator_filters}
\title{Read filters from wide format data}
\usage{
read_wide_indicator_filters(data, type)
}
\arguments{
\item{data}{Wide format data}

\item{type}{The type of data set}
}
\value{
Indicator filters
}
\description{
Expect input data with headers like
}
\details{
x, y, z, prevalence, art_cov

So a column for each separate indicator. For the data type we have look
for the list of possible indicators from the metadata. If that indicator
exists in the metadata return it as a possible filter with ID and name.
}
\keyword{internal}
