% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run the model}
\usage{
run_model(data, parameters)
}
\arguments{
\item{data}{Paths to different types of input data for the model.}

\item{parameters}{Any additional parameters for the model run.}
}
\value{
The job ID of the created model run.
}
\description{
This will prepare any data required by the model and call the model code.
}
\details{
Expecting input data to be a list of paths
list(
pjnz = "path/to/file",
shape = "path/to/file",
...
)
}
\keyword{internal}
