% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{stitch_options_template}
\alias{stitch_options_template}
\title{Stitch together separate sections of the options template}
\usage{
stitch_options_template(options_template)
}
\arguments{
\item{options_template}{List of separate options sections}
}
\value{
The stiched together options template
}
\description{
Stitch together separate sections of the options template
}
\keyword{internal}
