% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{to_redis_key}
\alias{to_redis_key}
\title{Convert a set of names to a key appropriate for storing in redis.}
\usage{
to_redis_key(name, ...)
}
\arguments{
\item{name}{Base name for redis key}

\item{...}{Any additional names for key}
}
\value{
The redis key name
}
\description{
Convert a set of names to a key appropriate for storing in redis.
}
\keyword{internal}
