% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{validate_json_schema}
\alias{validate_json_schema}
\title{Validate string of JSON against specified schema.}
\usage{
validate_json_schema(json, schema, query = NULL)
}
\arguments{
\item{json}{The JSON to validate.}

\item{schema}{Name of the schema to validate against.}
}
\value{
True if JSON adheres to schema.
}
\description{
This check will only be done if environmental variable VALIDATE_JSON_SCHEMAS
is set to true. If VALIDATE_JSON_SCHEMAS is not true then this always
returns TRUE.
}
\keyword{internal}
