% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_options.R
\name{verify_default}
\alias{verify_default}
\title{Select default value from hardcoded and input data and verify it}
\usage{
verify_default(name, controls, data_defaults, hardcoded_defaults)
}
\arguments{
\item{name}{Name of the control to verify}

\item{controls}{Named list of controls from the options JSON}

\item{merged_defaults}{List containg merge of options from data and
hardcoding in naomi.}
}
\value{
A named list of list containing any options and default values
for controls.
}
\description{
Verify that (if set) hardcoded default value is in the set of possible
options otherwise fallback to data default otherwise fallback to no
selection
}
\keyword{internal}
