% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Build and start the API}
\usage{
api(
  queue_id = NULL,
  workers = 2,
  results_dir = tempdir(),
  inputs_dir = NULL,
  log_level = "info"
)
}
\arguments{
\item{queue_id}{ID of an existing queue to connect to, creates a new one
if NULL}

\item{workers}{Number of workers to spawn}

\item{results_dir}{The dir for results to be saved to}

\item{inputs_dir}{THe directory where input files are stored}

\item{log_level}{The "lgr" log level to use}
}
\value{
Running API
}
\description{
Build and start the API
}
