% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{api}
\alias{api}
\title{Build and start the API}
\usage{
api(
  port = 8888,
  queue_id = NULL,
  workers = 2,
  results_dir = tempdir(),
  prerun_dir = NULL
)
}
\arguments{
\item{port}{Port for API}

\item{queue_id}{ID of an existing queue to connect to, creates a new one
if NULL}

\item{workers}{Number of workers to spawn}

\item{results_dir}{The dir for results to be saved to}

\item{prerun_dir}{The directory to store prerun results}
}
\value{
Running API
}
\description{
Build and start the API
}
