% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{do_validate_anc}
\alias{do_validate_anc}
\title{Validate ANC data file.}
\usage{
do_validate_anc(anc, shape, strict = TRUE)
}
\arguments{
\item{anc}{Path to input population file.}

\item{shape}{A file object (path, hash, filename) corresponding to
the input shape file.}

\item{strict}{If FALSE then run less stringent validation rules, used
for data exploration mode.}

\item{pjnz}{A file object (path, hash, filename) corresponding to
the input PJNZ file.}
}
\value{
An error if invalid.
}
\description{
Check that ANC data file can be read and return serialised data.
}
\keyword{internal}
