% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{do_validate_survey}
\alias{do_validate_survey}
\title{Validate survey data file.}
\usage{
do_validate_survey(survey, shape)
}
\arguments{
\item{survey}{Path to input survey file.}
}
\value{
An error if invalid.
}
\description{
Check that survey data file can be read and return serialised data.
}
\keyword{internal}
