% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{download_debug}
\alias{download_debug}
\title{Get debug output from naomi}
\usage{
download_debug(id, server = NULL, dest = tempfile(), verbose = TRUE)
}
\arguments{
\item{id}{The model run id to download. This will be printed at
the top of the stack trace.}

\item{server}{The url of the server.  The default is to use the
production naomi/hint/hintr instance. It is not possible to use
the staging instance.  You can change this if running locally.}

\item{dest}{The destination for the downloaded data.  The actual
data will be unpacked into a directory corresponding to the run
id within this, so it is safe to use a common directory.}

\item{verbose}{Add a progress bar}
}
\description{
Get debug output from naomi
}
