% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{endpoint_model_options}
\alias{endpoint_model_options}
\title{Get function to generate model options from Naomi template and input files}
\usage{
endpoint_model_options(req, res, shape, survey, programme = NULL, anc = NULL)
}
\arguments{
\item{req}{The request object.}

\item{res}{The response object.}

\item{shape}{The shape file.}

\item{survey}{The survey file.}

\item{programme}{The optional programme ART file.}

\item{anc}{The optional programme ANC file.}
}
\value{
Function to generate model options from input data.
}
\description{
This is used to build the model run options UI in the front end.
}
\keyword{internal}
