% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{hintr_response}
\alias{hintr_response}
\title{Format a hintr response and validate against schema.}
\usage{
hintr_response(value)
}
\arguments{
\item{value}{List containing an indication of success, any errors and the
value to return.}
}
\value{
Formatted hintr response.
}
\description{
Returns the status, any errors occured and the data if successful.
}
\keyword{internal}
