% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{hintr_response}
\alias{hintr_response}
\title{Format a hintr response and validate against schema.}
\usage{
hintr_response(value, schema, include_version = FALSE, as_json = TRUE)
}
\arguments{
\item{value}{List containing an indication of success, any errors and the
value to return.}

\item{schema}{The name of data subschema to validate response against.}

\item{include_version}{If TRUE the package version information is included
in the response.}

\item{as_json}{Logical, indicating if the response should be
converted into a json string}
}
\value{
Formatted hintr response.
}
\description{
Returns the status, any errors occured and the data if successful.
}
\keyword{internal}
