% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerun.R
\name{prerun_import}
\alias{prerun_import}
\title{Import prerun model results}
\usage{
prerun_import(prerun, path, model_output = "model-output.qs")
}
\arguments{
\item{prerun}{Path to the prerun model store}

\item{path}{Path to your results}

\item{model_output}{Path, within \code{path} to your model output qs}
}
\description{
Import prerun model results
}
