% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerun.R
\name{prerun_import}
\alias{prerun_import}
\title{Import prerun model results}
\usage{
prerun_import(
  prerun,
  path,
  output = "output.rds",
  spectrum = "spectrum.zip",
  summary = "summary.zip"
)
}
\arguments{
\item{prerun}{Path to the prerun model store}

\item{path}{Path to your results}

\item{output}{Path, within \code{path} to your output file}

\item{spectrum}{Path, within \code{path} to your spectrum file}

\item{summary}{Path, within \code{path} to your summary file}
}
\description{
Import prerun model results
}
