% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prerun.R
\name{prerun_push}
\alias{prerun_push}
\title{Push prerun model results to naomi.dide.ic.ac.uk}
\usage{
prerun_push(
  path,
  output = "output.rds",
  spectrum = "spectrum.zip",
  summary = "summary.zip"
)
}
\arguments{
\item{path}{Path to your results}

\item{output}{Path, within \code{path} to your output file}

\item{spectrum}{Path, within \code{path} to your spectrum file}

\item{summary}{Path, within \code{path} to your summary file}
}
\description{
Push prerun model results to naomi.dide.ic.ac.uk
}
